refactor: re-generate api client name using model directory name to prevent duplication
BREAKING CHANGE:
- classes were mass renamed. Refer to PR #79 for details.
