+++ src/api-clients/catalog-items-api-client.ts
@@ -1,9 +1,9 @@
-import { Configuration, DefaultApi } from '../api-models/finances-api-model'
+import { CatalogApi, Configuration } from '../api-models/catalog-items-api-model'
 import { ApiClientHelpers } from '../helpers'
 import { DEFAULT_API_BASE_PATH } from '../types'
 import { APIConfigurationParameters } from '../types/api-clients/api-configuration-parameters'
 
-export class DefaultApiClient extends DefaultApi {
+export class CatalogItemsApiClient extends CatalogApi {
   constructor(parameters: APIConfigurationParameters) {
     const axios = ApiClientHelpers.getAxiosInstance(parameters)
