+++ src/api-clients/fulfillment-inbound-api-client.ts
@@ -3,8 +3,7 @@ import { ApiClientHelpers } from '../helpers'
 import { DEFAULT_API_BASE_PATH } from '../types'
 import { APIConfigurationParameters } from '../types/api-clients/api-configuration-parameters'
 
-// TODO: duplicated API model with fba inbound
-export class FulfillmentFbaInboundApiClient extends FbaInboundApi {
+export class FulfillmentInboundApiClient extends FbaInboundApi {
   constructor(parameters: APIConfigurationParameters) {
     const axios = ApiClientHelpers.getAxiosInstance(parameters)
