+++ src/api-clients/index.ts
@@ -1,23 +1,23 @@
 export * from './aplus-content-api-client'
 export * from './authorization-api-client'
-export * from './catalog-api-client'
-export * from './fba-inbound-api-client'
+export * from './catalog-items-api-client'
+export * from './fba-inbound-eligibility-api-client'
 export * from './fba-inventory-api-client'
-export * from './small-and-light-api-client'
+export * from './fba-small-and-light-api-client'
 export * from './feeds-api-client'
-export * from './default-api-client'
-export * from './fulfillment-fba-inbound-api-client'
-export * from './fulfillment-fba-outbound-api-client'
+export * from './finances-api-client'
+export * from './fulfillment-inbound-api-client'
+export * from './fulfillment-outbound-api-client'
 export * from './merchant-fulfillment-api-client'
 export * from './messaging-api-client'
 export * from './notifications-api-client'
 export * from './orders-api-client'
-export * from './fees-api-client'
+export * from './product-fees-api-client'
 export * from './product-pricing-api-client'
 export * from './reports-api-client'
 export * from './sales-api-client'
 export * from './sellers-api-client'
-export * from './service-api-client'
+export * from './services-api-client'
 export * from './shipping-api-client'
 export * from './solicitations-api-client'
 export * from './tokens-api-client'
