+++ src/api-clients/product-fees-api-client.ts
@@ -3,7 +3,7 @@ import { ApiClientHelpers } from '../helpers'
 import { DEFAULT_API_BASE_PATH } from '../types'
 import { APIConfigurationParameters } from '../types/api-clients/api-configuration-parameters'
 
-export class FeesApiClient extends FeesApi {
+export class ProductFeesApiClient extends FeesApi {
   constructor(parameters: APIConfigurationParameters) {
     const axios = ApiClientHelpers.getAxiosInstance(parameters)
