+++ src/api-clients/services-api-client.ts
@@ -3,7 +3,7 @@ import { ApiClientHelpers } from '../helpers'
 import { DEFAULT_API_BASE_PATH } from '../types'
 import { APIConfigurationParameters } from '../types/api-clients/api-configuration-parameters'
 
-export class ServiceApiClient extends ServiceApi {
+export class ServicesApiClient extends ServiceApi {
   constructor(parameters: APIConfigurationParameters) {
     const axios = ApiClientHelpers.getAxiosInstance(parameters)
