+++ src/api-models/reports-api-model/api.ts
@@ -5,7 +5,7 @@
  * The Selling Partner API for Reports lets you retrieve and manage a variety of reports that can help selling partners manage their businesses.
  *
  * The version of the OpenAPI document: 2020-09-04
- *
+ * 
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
@@ -112,7 +112,7 @@ export interface CreateReportScheduleResult {
     reportScheduleId: string;
 }
 /**
- *
+ * 
  * @export
  * @interface CreateReportScheduleSpecification
  */
@@ -175,7 +175,7 @@ export enum CreateReportScheduleSpecificationPeriodEnum {
 }
 
 /**
- *
+ * 
  * @export
  * @interface CreateReportSpecification
  */
@@ -294,7 +294,7 @@ export interface GetReportSchedulesResponse {
  */
 export interface GetReportsResponse {
     /**
-     *
+     * 
      * @type {Array<Report>}
      * @memberof GetReportsResponse
      */
@@ -338,7 +338,7 @@ export interface ModelError {
     details?: string;
 }
 /**
- *
+ * 
  * @export
  * @interface Report
  */
@@ -424,7 +424,7 @@ export enum ReportProcessingStatusEnum {
 }
 
 /**
- *
+ * 
  * @export
  * @interface ReportDocument
  */
@@ -570,7 +570,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -603,7 +603,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -615,7 +615,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
         },
         /**
          * Creates a report.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0167 | 15 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportSpecification} body
+         * @param {CreateReportSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -635,7 +635,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             localVarHeaderParameter['Content-Type'] = 'application/json';
 
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
@@ -650,7 +650,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
         },
         /**
          * Creates a report schedule. If a report schedule with the same report type and marketplace IDs already exists, it will be cancelled and replaced with this one.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0222 | 10 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportScheduleSpecification} body
+         * @param {CreateReportScheduleSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -670,7 +670,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             localVarHeaderParameter['Content-Type'] = 'application/json';
 
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
@@ -706,7 +706,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -739,7 +739,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -772,7 +772,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             const localVarQueryParameter = {} as any;
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -808,7 +808,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             }
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -876,7 +876,7 @@ export const ReportsApiAxiosParamCreator = function (configuration?: Configurati
             }
 
 
-
+    
             setSearchParams(localVarUrlObj, localVarQueryParameter, options.query);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
@@ -918,7 +918,7 @@ export const ReportsApiFp = function(configuration?: Configuration) {
         },
         /**
          * Creates a report.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0167 | 15 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportSpecification} body
+         * @param {CreateReportSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -928,7 +928,7 @@ export const ReportsApiFp = function(configuration?: Configuration) {
         },
         /**
          * Creates a report schedule. If a report schedule with the same report type and marketplace IDs already exists, it will be cancelled and replaced with this one.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0222 | 10 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportScheduleSpecification} body
+         * @param {CreateReportScheduleSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -1022,7 +1022,7 @@ export const ReportsApiFactory = function (configuration?: Configuration, basePa
         },
         /**
          * Creates a report.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0167 | 15 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportSpecification} body
+         * @param {CreateReportSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -1031,7 +1031,7 @@ export const ReportsApiFactory = function (configuration?: Configuration, basePa
         },
         /**
          * Creates a report schedule. If a report schedule with the same report type and marketplace IDs already exists, it will be cancelled and replaced with this one.  **Usage Plan:**  | Rate (requests per second) | Burst | | ---- | ---- | | 0.0222 | 10 |  For more information, see \"Usage Plans and Rate Limits\" in the Selling Partner API documentation.
-         * @param {CreateReportScheduleSpecification} body
+         * @param {CreateReportScheduleSpecification} body 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
@@ -1127,7 +1127,7 @@ export interface ReportsApiCancelReportScheduleRequest {
  */
 export interface ReportsApiCreateReportRequest {
     /**
-     *
+     * 
      * @type {CreateReportSpecification}
      * @memberof ReportsApiCreateReport
      */
@@ -1141,7 +1141,7 @@ export interface ReportsApiCreateReportRequest {
  */
 export interface ReportsApiCreateReportScheduleRequest {
     /**
-     *
+     * 
      * @type {CreateReportScheduleSpecification}
      * @memberof ReportsApiCreateReportSchedule
      */
@@ -1366,3 +1366,5 @@ export class ReportsApi extends BaseAPI {
         return ReportsApiFp(this.configuration).getReports(requestParameters.reportTypes, requestParameters.processingStatuses, requestParameters.marketplaceIds, requestParameters.pageSize, requestParameters.createdSince, requestParameters.createdUntil, requestParameters.nextToken, options).then((request) => request(this.axios, this.basePath));
     }
 }
+
+
